Move CSS injection into utility helper Adjust app.json for Heroku PR deployments
